move logs_to_dot and show_logs graphviz to visualization
